pyproject.toml
[tool.poetry]
name = "flask-security-mwe"
version = "0.1.0"
description = ""
authors = ["andyfcx <dreamcastle.andy@gmail.com>"]
readme = "README.md"
packages = [{include = "flask_security_mwe"}]

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^2.3.2"
Flask-Security-Too = "^5.2.0"
SQLAlchemy = "^2.0.15"
psycopg2-binary = "^2.9.6"
gunicorn = "^20.1.0"
flask-sqlalchemy = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
